Create Exercise 2.8
Prompt the user to input an integer between 32 and 126, a float, a character, and a string, storing each into separate variables. Then, output those four values on a single line separated by a space. Extend to also output in reverse. Extend to convert the integer to a character by using the 'chr()' function, and output that character. Good luck! :)