Create Exercise 3.9
Prompt the user to enter two words and a number, storing each into separate variables. Then, output those three values on a single line separated by a space.  Output two passwords using a combination of the user input. Format the passwords as shown below. Output the length of each password (the number of characters in the strings). Good luck! :)